requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
MouseInfo==0.1.3
openai==2.6.1
PyAutoGUI==0.9.54
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.1
pytweening==1.2.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0